fix(indexer): add Dockerfile & container runtime definition for the indexer service
Closes #1327

- Add indexer/Dockerfile: multi-stage Rust build using cargo-chef for
  cached dependency compilation (planner -> builder -> runtime). The
  runtime stage is a slim Debian image with libssl-dev, libsqlite3-dev,
  ca-certificates and curl, runs as a non-root appuser (uid 10001),
  exposes port 3001 and ships a /health HEALTHCHECK via curl. The binary
  is installed to /usr/local/bin so the docker-compose bind-mount of
  ./indexer:/app never hides it.
- Add indexer/.dockerignore to keep the build context small.
- Fix the indexer crate so the Docker build can actually compile it:
  the crate did not build on main (stale Cargo.lock missing async-graphql
  entirely, plus async-graphql 7.x API drift, axum 0.7 -> 0.8, missing
  sqlx::FromRow derives, ledger type and missing imports).
- Update Cargo.toml/Cargo.lock to align axum 0.8, tower-http 0.6 and
  async-graphql 7.2 (with dataloader, apollo_persisted_queries and
  boxed-trait features).
- Fix quorum unanimous-state logic and extend the MockDb test helper so
  the crate tests pass.

+++ indexer/src/db/trait_.rs
@@ -2,17 +2,17 @@ use async_trait::async_trait;
 use anyhow::Result;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Serialize, Deserialize, Clone, sqlx::FromRow)]
 pub struct Event {
     pub id: String,
     pub contract_id: String,
-    pub ledger: u32,
+    pub ledger: i64,
     pub ledger_closed_at: String,
     pub event_type: String,
     pub data: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Serialize, Deserialize, Clone, sqlx::FromRow)]
 pub struct Quorum {
     pub id: String,
     pub quorum_type: String,
@@ -24,7 +24,7 @@ pub struct Quorum {
     pub expires_at: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Serialize, Deserialize, Clone, sqlx::FromRow)]
 pub struct Vote {
     pub id: String,
     pub quorum_id: String,
@@ -34,7 +34,7 @@ pub struct Vote {
     pub timestamp: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Serialize, Deserialize, Clone, sqlx::FromRow)]
 pub struct Oracle {
     pub id: String,
     pub name: String,
@@ -42,7 +42,7 @@ pub struct Oracle {
     pub active: bool,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Serialize, Deserialize, Clone, sqlx::FromRow)]
 pub struct AuditEntry {
     pub id: String,
     pub event_type: String,
